+++ packages/atomic/src/components/search/atomic-search-interface/atomic-search-interface.tsx
@@ -12,6 +12,8 @@ import {
   loadQueryActions,
   EcommerceDefaultFieldsToInclude,
   loadFieldActions,
+  getOrganizationEndpoints as getOrganizationEndpointsHeadless,
+  PlatformEnvironment,
 } from '@coveo/headless';
 import {
   Component,
@@ -278,7 +280,7 @@ export class AtomicSearchInterface
   }
 
   /**
-   * Initializes the connection with the headless search engine using options for `accessToken` (required), `organizationId` (required), `renewAccessToken`, and `platformUrl`.
+   * Initializes the connection with the headless search engine using options for accessToken (required), organizationId (required), renewAccessToken, organizationEndpoints (recommended), and platformUrl (deprecated).
    */
   @Method() public initialize(options: InitializationOptions) {
     return this.internalInitialization(() => this.initEngine(options));
@@ -336,6 +338,18 @@ export class AtomicSearchInterface
     const {value, enableQuerySyntax, analytics} = standaloneSearchBoxData;
     this.engine!.dispatch(updateQuery({q: value, enableQuerySyntax}));
     this.engine.executeFirstSearchAfterStandaloneSearchBoxRedirect(analytics);
+  }
+
+  /**
+   * Returns the unique, organization-specific endpoint(s)
+   * @param {string} organizationId
+   * @param {'prod'|'hipaa'|'staging'|'dev'} [env=Prod]
+   */
+  @Method() public async getOrganizationEndpoints(
+    organizationId: string,
+    env: PlatformEnvironment = 'prod'
+  ) {
+    return getOrganizationEndpointsHeadless(organizationId, env);
   }
 
   public get bindings(): Bindings {